ktest.pl: Fix the logic for truncating the size of the log
Git-commit: 2902e302991a4dacd155658bcb3053e4d6b7d4a0
 file for email
References: bsn#19
Patch-mainline: v5.10.2

commit 170f4869e66275f498ae4736106fb54c0fdcd036 upstream.

The logic for truncating the log file for emailing based on the
MAIL_MAX_SIZE option is confusing and incorrect. Simplify it and have the
tail of the log file truncated to the max size specified in the config.

Cc: stable@vger.kernel.org
Fixes: 855d8abd2e8ff ("ktest.pl: Change the logic to control the size of the log file emailed")
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.2-0003-ktest.pl-Fix-the-logic-for-truncating-the-size-of-th.patch