scsi: ufs-pci: Ensure UFS device is in PowerDown mode for
Git-commit: 46e4355836104706853819244adf21ae1e3cd621
 suspend-to-disk ->poweroff()
References: bsn#19
Patch-mainline: v5.10.7

[ Upstream commit af423534d2de86cd0db729a5ac41f056ca8717de ]

The expectation for suspend-to-disk is that devices will be powered-off, so
the UFS device should be put in PowerDown mode. If spm_lvl is not 5, then
that will not happen. Change the pm callbacks to force spm_lvl 5 for
suspend-to-disk poweroff.

Link: https://lore.kernel.org/r/20201207083120.26732-3-adrian.hunter@intel.com
Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.7-0054-scsi-ufs-pci-Ensure-UFS-device-is-in-PowerDown-mode-.patch