mptcp: add port support for ADD_ADDR suboption writing
Git-commit: 22fb85ffaefb80a22c815008a500273b3f61bba3
Message-Id: <22fb85ffaefb80a22c815008a500273b3f61bba3.1637075074.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.11-rc1

In rfc8684, the length of ADD_ADDR suboption with IPv4 address and port
is 18 octets, but mptcp_write_options is 32-bit aligned, so we need to
pad it to 20 octets. All the other port related option lengths need to
be added up 2 octets similarly.

This patch added a new field 'port' in mptcp_out_options. When this
field is set with a port number, we need to add up 4 octets for the
ADD_ADDR suboption, and put the port number into the suboption.

Signed-off-by: Geliang Tang <geliangtang@gmail.com>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.11-rc1-mptcp-add-port-support-for-ADD_ADDR-suboption-writin.patch