block/rnbd: Set write-back cache and fua same to the target
From 512c781fd28cb401ee9f2843e32bf4640732c671 Mon Sep 17 00:00:00 2001
 device
Git-commit: 512c781fd28cb401ee9f2843e32bf4640732c671
Patch-mainline: v5.11-rc1
References: bsn#1

The rnbd-client always sets the write-back cache and fua attributes
of the rnbd device queue regardless of the target device on the server.
That generates IO hang issue when the target device does not
support both of write-back cacne and fua.

This patch adds more fields for the cache policy and fua into the
device opening message. The rnbd-server sends the information
if the target device supports the write-back cache and fua
and rnbd-client recevives it and set the device queue accordingly.

Signed-off-by: Gioh Kim <gi-oh.kim@cloud.ionos.com>
[jwang: some minor change, rename a few varables, remove unrelated comments.]
Signed-off-by: Jack Wang <jinpu.wang@cloud.ionos.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.suse/v5.11-rc1-block-rnbd-Set-write-back-cache-and-fua-same-to-the-.patch